+++ CHANGELOG.md
@@ -8,6 +8,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - A namespace prefix can be added to all metrics.
 - An optional status monitoring API, serving JSON over HTTP
 
+### Bugfixes
+
+- If response code from OpenTSDB is 400, don't retry sending since this means we
+  sent bad data.
+  
+
 ## [1.3.1](https://github.com/OpenTSDB/tcollector/issues?utf8=%E2%9C%93&q=milestone%3A1.3.1+)
 ### Collectors Added
 - docker.py - Pulls metrics from a local Docker instance, tries /var/run/docker.sock, then localhost API